Refactor MainForm responsibilities into partial classes
